created a folder to host various components and it styles
