+++ src/scss/index.scss
@@ -1,8 +1,8 @@
 @import 'variables';
 @import 'globals';
 @import 'components/index';
-@import 'section/index';
-@import 'section/sponsors';
+@import 'section/LandingPageStyles/index';
+ @import 'section/LandingPageStyles/sponsors';
