+++ src/scss/section/LandingPageStyles/_sponsors.scss
