+++ src/scss/section/LandingPageStyles/_what_do.scss
