+++ src/section/LandingPage/SecondExperience.tsx
@@ -1,15 +1,16 @@
 import React from 'react'
-import InspoImage from '../../public/assets/images/artsszone.png';
+import InspoImage from '../../../public/assets/images/artsszone.png';
 import Image from 'next/image';
 
-function ThirdExperience() {
+
+function SecondExperience() {
   return (
-    <div className='first-experience-container'>
-      <div className="first-experience-container-elements">
-         <div className="first-experience-container-elements-image">
+<div className='second-experience-container'>
+      <div className="second-experience-container-elements">
+         <div className="second-experience-container-elements-image">
             <Image src={InspoImage} alt="expo-1"/>
          </div>
-         <div className="first-experience-container-elements-text">
+         <div className="second-experience-container-elements-text">
             <h3>Lorem ipsum dolor sit amet.</h3>
             < div className='p-tag'>
             <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Maiores sed illo aut debitis, perspiciatis necessitatibus odit velit dicta rem a ad alias et sequi voluptas voluptatem ipsa nulla saepe totam non quo explicabo incidunt! Ullam mollitia molestias dolorum 
@@ -17,7 +18,7 @@ function ThirdExperience() {
             </div>    
 
             <div className="button">
-                <button>get to know</button>
+                <button>tell me more</button>
             </div>
 
          </div>
@@ -26,4 +27,4 @@ function ThirdExperience() {
   )
 }
 
-export default ThirdExperience
+export default SecondExperience
